Move MaterialUtils and MaterialStore to their own files. Renamed TransferableUtils to TransportUtils. Updated Typescript definitions.
